pyproject.toml
[tool.poetry]
name = "navi"
version = "0.0.1"
description = "Personal assistant"
authors = ["Andrew Yoon <andrewy563@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
python-dotenv = "^0.19.2"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.1.0"
isort = "^5.10.1"

[tool.poetry.scripts]
start = "navi.bot:start"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length=100
skip_gitignore=true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
